PR: 알림, 마스터 테이블, 어댑터 등 추가 (#25)
* feat: POST 마스터 테이블 추가
Project 연결 완료

* feat: Create 병합

* feat: 게시글 조회 메서드 추가
ID를 넣으면 공지와 프로젝트가 각각 반환됨

* feat: 게시글 업데이트 메서드 추가
서비스 내 주석 추가

* feat: 마스터 댓글 CRUD 추가

* feat: 게시글 ID로 댓글 조회하는 기능 추가

* feat: 타입별 구분하여 게시글 목록 읽는 메서드 추가

* chore: Compact 기능 추가
true 설정 시 타입별 게시글 읽기 시 제목과 내용, id만 반환하도록 리소스 최적화가 이루어집니다.

* refact: sse 패키지 위치 변경

* feat: Linked-id 반환 메서드 추가

* feat: 알림센터 테이블 추가 및 읽기 컨트롤 추가

* feat: 읽은 알림, 안 읽은 알림 따로 보기 기능 추가

* feat: 최대 Noti 개수 유지하도록 설정
우선은 20개까지만.

* feat: Noti 삭제 기능 추가

* chore: maxNoti... 접근 지정자 및 네이밍 규칙 수정

* chore: Comments 수정 메서드 Patch로 변경

* feat: 알림 전체 삭제 API 추가

* feat: 알림 전체 읽기 API 추가

* chore: 게시글 입력 시 작성자 및 시간표시 컬럼 추가

* feat: 작성 글 목록 조회 고도화
제목까지 표시하도록 개발

* feat: 작성 댓글 목록 조회 고도화
작성 내용과 POST ID까지 반환

* feat: 스크랩 (찜) CRUD 테이블 생성

* feat: 스크랩 조회 Mypage 연결

+++ src/main/kotlin/io/sprout/api/mypage/repository/DummyPostParticipantRepository.kt
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface DummyPostParticipantRepository : JpaRepository<DummyPostParticipant, Int> {
-    fun findAllByUserId(userId: Int): List<DummyPostParticipant>
+interface DummyPostParticipantRepository : JpaRepository<DummyPostParticipant, Long> {
+    fun findAllByUserId(userId: Long): List<DummyPostParticipant>
 
-    fun deleteByPostparticipantidAndUserId(postparticipantid: Int, userId: Int): Int
+    fun deleteByPostparticipantidAndUserId(postparticipantid: Int, userId: Long): Int
 }
